% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{aric_coef}
\alias{aric_coef}
\title{ARIC Coefficients}
\format{
A named vector with coefficients for each protein's first 2 FPC
}
\usage{
aric_coef
}
\description{
This dataset contains the coefficients or each protein's first 2 FPC trained from 4684 proteins from the ARIC study
}
\details{
The coefficients were extracted from the trained models and can be used for directly calculating LPAI.
}
\keyword{datasets}
