% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ARIC_trained_models}
\alias{ARIC_trained_models}
\title{ARIC Trained Models}
\format{
A list containing trained FPCA models and training FPC scores for each protein:
}
\usage{
ARIC_trained_models
}
\description{
This data contains the trained FPCA model for each protein based on the ARIC study data.
}
\details{
The models were trained using the FPCA method on ARIC study data.
}
\keyword{datasets}
