% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{example.test.data}
\alias{example.test.data}
\title{Example Test data}
\format{
a dataframe
}
\usage{
example.test.data
}
\description{
This data contains example test data with columns id, age and proteins
}
\keyword{datasets}
