% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{example.train.data}
\alias{example.train.data}
\title{Example training data}
\format{
a dataframe
}
\usage{
example.train.data
}
\description{
This data contains example training data with columns id, age and proteins
}
\keyword{datasets}
