% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LPAI.R
\name{get_LPAI}
\alias{get_LPAI}
\title{Calculate LPAI Using FPC Scores and Cox Coefficients}
\usage{
get_LPAI(test_fpc_scores, cox_coefs)
}
\arguments{
\item{test_fpc_scores}{A data frame containing FPC scores for the test data. The first column must be \code{id}.}

\item{cox_coefs}{A named numeric vector of coefficients from the trained Cox model.}
}
\value{
A data frame with \code{id} and the computed risk score (\code{LPAI}).
}
\description{
This function applies Cox regression coefficients to compute the LPAI (risk score) for each individual.
}
\examples{
\dontrun{
#Load example data
data(ARIC_trained_models)
data(example.test.data)
data(aric_coef)
#Predict FPC scores
test_fpc_scores <- predict_protein_FPCscores(
test_data = example.test.data,
trained_models = ARIC_trained_models,
protein_indices = c(3, 4),
nFPCs = 2,Seqid = T
)
lpai<-get_LPAI(test_fpc_scores, aric_coef)
print(lpai)

}

}
