% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_protein_FPCscores.R
\name{predict_protein_FPCscores}
\alias{predict_protein_FPCscores}
\title{Predict Functional Principal Component (FPC) Scores for Test Data}
\usage{
predict_protein_FPCscores(
  test_data,
  trained_models = ARIC_trained_models,
  protein_indices,
  nFPCs = 2,
  Seqid = F
)
}
\arguments{
\item{test_data}{A data frame containing test data with columns \code{id}, \code{age}, and protein data.}

\item{trained_models}{A named list of trained FPCA models. Default is \code{ARIC_trained_models}.}

\item{protein_indices}{A numeric vector specifying the column indices of protein data in \code{test_data}.}

\item{nFPCs}{An integer specifying the number of FPC scores to extract. Default is 2.}

\item{Seqid}{Logical. If \code{TRUE}, rename user columns to UniProt IDs. Default is \code{FALSE}.}
}
\value{
A data frame with rows corresponding to test samples, containing \code{id} and FPC scores for each protein.
The FPC scores are named as \code{protein_name_FPC1}, \code{protein_name_FPC2}, etc.
}
\description{
This function predicts FPC scores for test data using trained FPCA models.
It validates input data, handles duplicate observations, filters out age values outside the training range,
and returns the requested FPC scores for each protein. The function will print messages to the console during execution.
}
\examples{
\dontrun{
# Load example data
data(ARIC_trained_models)
data(example.test.data)

# Predict FPC scores
test_fpc_scores <- predict_protein_FPCscores(
 test_data = example.test.data,
 trained_models = ARIC_trained_models,
 protein_indices = c(3, 4),
 nFPCs = 2,Seqid = T
)
print(test_fpc_scores)
}
}
