% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for FPCA (Using Protein Index)}
\usage{
prepare_data(data, protein_index)
}
\arguments{
\item{data}{A data frame containing training data with columns \code{id}, \code{age}, and protein data.}

\item{protein_index}{An integer specifying the index of the protein column to be converted.}
}
\value{
A wide-format data frame where rows represent unique \code{age} values and columns represent protein values for each \code{id}.
}
\description{
This function processes data to convert it into a wide format for a specified protein column by its index.
The resulting wide-format data has \code{age} as rows and protein values for each \code{id} as columns.
}
